feat(home): fetch most-viewed and best-selling product lists from server

+++ lib/domain/entities/product_entity.dart
@@ -11,7 +11,7 @@ class ProductEntity {
   int quantity;
   String thumbnail;
   int realPrice;
-  int discountPercent;
+  double discountPercent;
 
   ProductEntity(
     this.category,
